# Speed up pointcloud output by colorizing depth images instead of depthifying color images