## Fixes #
Partially addresses #56 

Hopefully makes @ooeygui and @RoseFlunder happier with the default behavior of the driver. :)

### Description of the changes:
The default behavior for the RGB colored pointcloud was to take the depth image and project it *into* the RGB coordinate frame (and image size) and publish the pointcloud in the RGB TF frame. While this makes sense for some applications (i.e., texturing meshes) as it gives you the biggest possible pointcloud, it also, well, gives you a large pointcloud that takes a while to compute.
And I would argue in the vast majority of cases, the correct thing to do is to colorize the depth cloud instead of depth-ifying the color image. This generally leads to a smaller pointcloud size, and also *is technically correct* from a sensor integration point of view.

You can imagine if you project these pointclouds into a volumetric map, having the sensor TF frame be the RGB frame rather than the depth is incorrect, as the measurements are actually made from the depth frame. Similarly, projecting the depth into the RGB frame creates overconfidence in the depth measurements as then interpolated depth is treated as real sensor measurements.

- Add `point_cloud_in_depth_frame` parameter which **now defaults to true**, defaulting to the new behavior. This parameter makes the colorized pointcloud be published in the depth coordinate frame, rather than the RGB image frame.
- This speeds up the pointcloud publishing significantly, by publishing a smaller pointcloud (see analysis below).

<!-- Please check off the appropriate boxes with [x] before submitting your pull request -->
### Before submitting a Pull Request:
- [x] I reviewed [CONTRIBUTING.md](https://github.com/microsoft/Azure_Kinect_ROS_Driver/blob/master/CONTRIBUTING.md)
- [x] I [built my changes](https://github.com/microsoft/Azure_Kinect_ROS_Driver/blob/master/docs/building.md) locally
- [x] I tested my changes with a device

### I tested changes on: <!-- it's not required to have tested both, just indicate which one you tried -->
- [x] Windows
- [x] Linux


<!-- Specify how you tested your changes (i.e. manual/ad-hoc testing, automated testing, new automated tests added)-->
# Timing analysis
Attempting to address exactly the issue in #56, which is that the pointcloud output cannot run at 30 Hz if the RGB color is used.

***Settings***: WFOV_2x2BINNED, 1536P, point_cloud = true, rgb_point_cloud = true, point_cloud_in_depth_frame = **false** (old behavior, pre-PR):
```
rostopic hz /points2
subscribed to [/points2]
average rate: 9.198
	min: 0.084s max: 0.118s std dev: 0.01112s window: 7
```
***Settings***: WFOV_2x2BINNED, 1536P, point_cloud = true, rgb_point_cloud = true, point_cloud_in_depth_frame = **true** (new behavior):

```
rostopic hz /points2
subscribed to [/points2]
average rate: 31.655
	min: 0.007s max: 0.034s std dev: 0.00590s window: 26
```

The table below is a more comprehensive analysis of the time it takes to build and publish a pointcloud to ROS.
**Depth in Depth**: point_cloud = true, rgb_point_cloud = false, point_cloud_in_depth_frame = false
**RGB in Depth**: point_cloud = true, rgb_point_cloud = true, point_cloud_in_depth_frame = true (new)
**RGB in RGB**: point_cloud = true, rgb_point_cloud = true, point_cloud_in_depth_frame = false (old)

<table><tr><th>Depth Mode</th><th colspan="2">WFOV Unbinned</th><th colspan="2">WFOV 2x2 Binned</th></tr><tr><td></td><td>720P</td><td>1536P</td><td>720P</td><td>1536P</td></tr><tr><td>Depth in Depth</td><td>10 ms</td><td>10 ms</td><td>10 ms</td><td>10 ms</td></tr><tr><td>RGB in Depth</td><td>33 ms</td><td>40 ms</td><td>9 ms</td><td>10 ms</td></tr><tr><td>RGB in RGB</td><td>21 ms</td><td>75 ms</td><td>21 ms</td><td>70 ms</td></tr></table>

As can be seen, the new setting is about 2x faster in all cases except 720p and WFOV Unbinned, since the RGB image actually has a *smaller* resolution that the depth image and projecting into the RGB space creates a smaller pointcloud.

# Screenshots
## Depth in Depth
![Screenshot from 2019-10-18 15-53-56](https://user-images.githubusercontent.com/5616392/67112590-ded0cf80-f1d7-11e9-97be-834327d84bfa.png)

## RGB in Depth
### (small point size)
![Screenshot from 2019-10-18 15-54-29](https://user-images.githubusercontent.com/5616392/67112654-06c03300-f1d8-11e9-96c1-234e6a5d3da3.png)

### (large point size)
![Screenshot from 2019-10-18 15-55-46](https://user-images.githubusercontent.com/5616392/67112691-1a6b9980-f1d8-11e9-9569-8411399276ad.png)


## RGB in RGB
![Screenshot from 2019-10-18 15-55-04](https://user-images.githubusercontent.com/5616392/67112673-117ac800-f1d8-11e9-9397-2c47ddbde015.png)