@RoseFlunder 2019-10-18
Thank you for your comparisons.
I switched myself to the "color to depth" transformation in my fork because of performance reasons.

But we had some other related issue before about "color to depth" and @skalldri argumented that he prefers "depth to color".
Here is his opinion:
https://github.com/microsoft/Azure_Kinect_ROS_Driver/issues/63#issuecomment-524957010

Just wanted to mention this. Nevertheless I am all in for an option to let the user decide which transformation is chosen for point cloud generation.