@helenol 2019-10-18
@RoseFlunder thanks for being such an expert on this driver! :)
Interesting discussion there!
But I agree letting users decide is the better way, and I am pretty sure that rgb in depth is the _standard_ way (see depth_image_proc: https://github.com/ros-perception/image_pipeline/blob/melodic/depth_image_proc/src/nodelets/point_cloud_xyzrgb.cpp  and other depthcam drivers) and makes more sense from a sensor fusion point of view as well.