@RoseFlunder 2019-10-18
I am glad to help when I can.
I am no expert in robotics or computer vision but at the moment I try to learn what I can about working with the Azure Kinect devices because I am going to use them in my master thesis project.
The department at my university has a multi kinect system in an ROS environment and just switched from the previous Kinect v2 to the Azure Kinects.
I shall be using them to implement a body tracking data fusion algorithm with some constraints like "works with at least 4 persons" and "persons can be partially occluded for some of the cameras" etc.
So my starting point was this package.