@skalldri 2019-10-18
This seems like the wrong solution, for two reasons.

First, as RoseFlunder mentioned, I've previously articulated why using RGB-to-depth results in incorrect RGB point clouds. To re-iterate: we cannot know which RGB pixels are observed only by the RGB camera and not by the depth camera (or vice versa), so when colorizing the depth map with an RGB frame you will incorrectly colorize depth pixels that the RGB camera did not observe. This results in "color bleed" from objects observed by the RGB camera onto depth pixels not associated with that object. This effect is more pronounced with objects closer to the camera. The Azure Kinect SDK documentation suggest using depth-to-RGB for this exact reason.

@helenol, you say:

> But I agree letting users decide is the better way, and I am pretty sure that rgb in depth is the standard way (see depth_image_proc: https://github.com/ros-perception/image_pipeline/blob/melodic/depth_image_proc/src/nodelets/point_cloud_xyzrgb.cpp and other depthcam drivers) and makes more sense from a sensor fusion point of view as well.

This isn't correct. This is why the [depth_image_proc/register](http://wiki.ros.org/depth_image_proc?distro=melodic#depth_image_proc.2BAC8-register) nodelet exists: you _must_ re-project depth images into the RGB co-ordinate frame before running them through depth_image_proc/point_cloud_xyzrgb. See the [depth_image_proc overview](http://wiki.ros.org/depth_image_proc?distro=melodic#Overview) which states (emphasis mine):

> depth_image_proc provides basic processing for depth images, much as image_proc does for traditional 2D images. The two packages are complementary; **for example, you can (and should!) rectify your depth image before converting it to a point cloud.**

You can even see a comment that implies this in the [point_cloud_xyzrgb code](https://github.com/ros-perception/image_pipeline/blob/d4bb3f00c3cac994190b828bf7d27602900ea90a/depth_image_proc/src/nodelets/point_cloud_xyzrgb.cpp#L162-L168). The original Kinect for Xbox One ROS Driver [does this as well](https://github.com/code-iai/iai_kinect2/tree/master/kinect2_registration)).

You also say: 

> You can imagine if you project these pointclouds into a volumetric map, having the sensor TF frame be the RGB frame rather than the depth is incorrect, as the measurements are actually made from the depth frame.

I agree with your statement if we were talking about un-colorized point clouds, and indeed you can see that when publishing un-colorized point clouds they are published with the depth camera TF frame. 

Fundamentally, the Azure Kinect cannot capture RGB point clouds. It captures sufficient data that allows us to _reconstruct_ an RGB point cloud, but RGB point clouds are not a native data format of the Azure Kinect. You claim it is wrong to publish this reconstructed point cloud in the RGB frame, since the depth data comes from the depth frame. But conversely, you could also claim it's wrong to publish it in the depth frame, since the RGB data that colorized each pixel came from the RGB frame. In both cases, some of the data is being published in the "wrong" frame from the "wrong" perspective.

Since either publishing frame (depth or RGB) can be considered the "wrong" frame, we should instead consider the accuracy of re-projection into either frame. Re-projecting depth to RGB is easy, since each depth pixel represents a point in 3D space. Producing a new depth image based on these 3D points can be done with extremely high accuracy, taking into account occlusions which may occur in the new co-ordinate frame. Re-projecting RGB to depth is not easy: there is no Z-depth associated with the pixels, so the re-projection cannot account for new occlusions that might occur when observing the scene from a new perspective.

It is therefore reasonable to publish this data from the perspective of the RGB camera, since it uses the highest accuracy re-projection method to preserve the image quality. 

Second, the point of opening this bug was to indicate that there is some significant inefficiency in the `getRgbPointCloud()` function. Publishing a point cloud with less pixels does not solve the fact that this function should have no problem rendering an RGB point cloud in significantly less than 30ms on pretty much any machine. RoseFlunder's previous investigations into timing showed that the bulk of that time was not spent in Azure Kinect SDK functions, so the source of the slowdown must lie in this driver.