@RoseFlunder 2019-10-18
@skalldri Thanks for the long explanation. I understand that the color of the point cloud pixels is not guaranteed to be correct and there will be color spill when using "color to depth", but when I mostly care about accurate positions (x,y,z) of the pixels and the color is just some "nice to have visual pleasing for the eyes" I could use or did I understand it wrong?

Nevertheless it would be fantastic if the time to transfer the SDK pointcloud data into a PointCloud2 message could be done faster. If it would run with 30 FPS (<33ms) with 1536P and "depth to color" it would be perfect.